✨♻️RM#4407-3: Improving overall code structure
What: add specific handler modules, one for os related tasks and one
for shell related tasks. Add one module to list all log messages used
in the program. Fix an issue in the use of copybooks in dataset
migration.
Why: improving the overall project to ease the maintenance of the code
and the introduction of future features.